+++ custom/justech_managed_services/__init__.py
@@ -1,6 +1,7 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0.html)
 from . import controllers
 from . import models
+from . import wizard
 from .hooks import post_init_hook
 
 __all__ = ["post_init_hook"]