+++ custom/justech_managed_services/models/__init__.py
@@ -1,4 +1,5 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0.html)
+from . import form_catalog
 from . import managed_service
 from . import managed_service_assessment
 from . import res_partner